Clean up code and entry points
Make use of helpers within the module to clean up primary entry
points, e.g. get_mets encapsulaes four or five other functions
that can all be accessed via those helpers. This should make for
more discrete units we can test.
